requirements.txt
asgiref==3.2.7
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2
django-import-export==2.0.2
django-restframework==0.0.1
django-widget-tweaks==1.4.8
et-xmlfile==1.0.1
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.3
pytz==2019.3
PyYAML==5.3.1
sqlparse==0.3.1
tablib==1.1.0
xlrd==1.2.0
xlwt==1.3.0
djangorestframework==3.11.0
pillow

whitenoise
six
s3transfer==0.2.1
gunicorn==20.0.4
django-heroku==0.3.1
boto3
django-storages==1.8
Jinja2==2.10.3
django-cors-headers
django-rest-auth
django-allauth
aioredis==1.2.0
asgiref==2.3.2
astroid==2.0.4
async-timeout==3.0.1
attrs==18.2.0
autobahn==18.10.1
Automat==0.7.0
certifi==2018.10.15
channels==2.1.5
channels-redis==2.3.1
chardet==3.0.4
constantly==15.1.0
daphne==2.2.2
defusedxml==0.5.0
django-allauth==0.38.0
django-cors-headers==2.4.0
django-rest-auth==0.9.3
django-rest-framework==0.1.0
djangorestframework==3.9.0
gunicorn==19.7.1
hiredis==0.2.0
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
msgpack==0.5.6
oauthlib==2.1.0
psycopg2-binary==2.7.5
PyHamcrest==1.9.0
pylint==2.1.1
python3-openid==3.1.0
pytz==2018.7
requests==2.20.1
requests-oauthlib==1.0.0
six==1.11.0
sqlparse==0.3.0
Twisted==18.9.0
txaio==18.8.1
typed-ast==1.1.0
urllib3==1.25.2
whitenoise==3.3.1
wrapt==1.10.11
zope.interface==4.6.0
